WIP: implemented proper hashable classes.
Using oid for all comparisons and hashing because it's less mutable than
the name attribute. According to Python docs the value used for
comparison and hashing must be immutable. If it changes then the
position in any hash table will be incorrect.

This needs extensive testing and is likely a breaking change because I
do not know why original author used name attribute.